Moved article_recommendation to interfaces
