+++ sciety_labs/providers/async_providers/opensearch/providers.py
@@ -7,11 +7,11 @@ import numpy.typing as npt
 
 import opensearchpy
 
-from sciety_labs.providers.article_recommendation import (
+from sciety_labs.providers.interfaces.article_recommendation import (
     ArticleRecommendationFilterParameters,
     ArticleRecommendationList
 )
-from sciety_labs.providers.article_recommendation import (
+from sciety_labs.providers.interfaces.article_recommendation import (
     AsyncSingleArticleRecommendationProvider
 )
 from sciety_labs.providers.async_providers.crossref.providers import (
